fix issues after deployment

+++ run.sh
@@ -11,7 +11,7 @@ delete_image() {
   read -a images <<<"$*"
 
   for image_name in "${images[@]}"; do
-    printf "[image] '%s' -> " "$image_name"
+    echo "[image] '%s' -> " "$image_name"
     docker rmi "$image_name"
   done
 
@@ -20,7 +20,7 @@ delete_image() {
 check_nginx_volume() {
   if ! docker volume ls -q | grep nginx_conf; then
     docker volume create --name=nginx_conf
-    printf "[INFO] nginx_conf was volume created "
+    echo "[INFO] nginx_conf was volume created "
   fi
 }
 
@@ -39,30 +39,47 @@ start() {
   docker compose logs -f
 }
 
-while getopts rdsh OPTION; do
-  args_count="$#"
-  case "$OPTION" in
-  r)
-    if [ $args_count -gt 1 ]; then
-      shift
-      delete_image "$*"
-    else
-      echo "[ERROR] You should write docker images names."
-    fi
-    ;;
-  d)
-    docker compose down
-    echo "[INFO] docker compose stopped, container removed."
-    ;;
-  s)
-    start
-    ;;
-  h)
-    help
-    ;;
-  ?)
-    echo "[ERROR] Wrong flag. Script usage: [-h help] [-r somevalues] [-s] [-u]"
-    exit 1
-    ;;
-  esac
-done
+args_count="$#"
+if [ $args_count == 0 ]; then
+  echo "[INFO] Start docker compose..."
+  start
+else
+
+  while getopts rdshia OPTION; do
+
+
+    case "$OPTION" in
+    r)
+      if [ $args_count -gt 1 ]; then
+        shift
+        delete_image "$*"
+      else
+        echo "[ERROR] You should write docker images names."
+      fi
+      ;;
+    d)
+      docker compose down
+      echo "[INFO] Docker compose stopped, containers removed."
+      ;;
+    a)
+      docker rmi $(docker images -q)
+      echo "[INFO] All docker images $pattern are deleted"
+      ;;
+    i)
+      pattern=$2
+      docker rmi $(docker images | grep $pattern)
+      echo "[INFO] All docker images '$pattern' are deleted"
+      ;;
+    h)
+      help
+      ;;
+    ?)
+      echo "[ERROR] Wrong flag. Script usage: [-h help] [-r somevalues] [-d] [-s]"
+      exit 1
+      ;;
+    *)
+
+    esac
+  done
+
+fi
